Move `NewButton` to `DataAppActionPanel`
